- WIP there is a problem with LiteDb and windows right now. Its can give have errors while try to connect to the DB even if its shared
-  **fixxed** WIP important to fix map choose view with the game modes cause if you save the server settings the maps will be overwritten with this set there.

+++ Services/PavlovServerService.cs
@@ -28,7 +28,7 @@ namespace PavlovRconWebserver.Services
             return _liteDb.LiteDatabase.GetCollection<PavlovServer>("PavlovServer").Include(x=>x.SshServer)
                 .Find(x=>x.SshServer.Id==sshServerId).ToList();
         }
-        public async Task<PavlovServer> FindOne(long id)
+        public async Task<PavlovServer> FindOne(int id)
         {
             var pavlovServer =  _liteDb.LiteDatabase.GetCollection<PavlovServer>("PavlovServer").Include(x=>x.SshServer)
                 .Find(x => x.Id == id).FirstOrDefault();
